LPS-148475 Added Table columns validation for non admin users
