+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/containers/TeamMembersTable/utils/getColumnsByUserAccess.js
@@ -1,0 +1,26 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ */
+
+import {COLUMNS} from './constants';
+
+const OPTIONS_COLUMN_NAME = 'options';
+
+export function getColumnsByUserAccess(hasAdminAccess) {
+	if (!hasAdminAccess) {
+		const ColumnsWithoutOptions = COLUMNS.filter(
+			(column) => column.accessor !== OPTIONS_COLUMN_NAME
+		);
+
+		return ColumnsWithoutOptions;
+	}
+
+	return COLUMNS;
+}
